analytics/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.1.3
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
